app/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
PyPDF2>=3.0.0
opik>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
langgraph>=0.0.1
python-dotenv>=1.0.0
httpx>=0.25.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
langchain-text-splitters>=0.0.1
numpy>=1.24.0
pywebview>=5.0.0
requests>=2.31.0
